разбил верстку на отдельные модули и импотирвал в main.js
