+++ src/view/board.js
@@ -1,0 +1,8 @@
+const createBoardTemplate = () => {
+    return (
+        `<section class="board container">
+          </section>`
+    );
+};
+
+export { createBoardTemplate };
