+++ src/view/loading-message.js
@@ -1,0 +1,9 @@
+const createLoadingMessageTemplate = () => {
+    return (
+        `<p class="board__no-tasks">
+              Loading...
+          </p>`
+    );
+};
+
+export { createLoadingMessageTemplate };
