+++ src/view/no-tasks-message.js
@@ -1,0 +1,9 @@
+const createNoTasksMessageTemplate = () => {
+    return (
+        `<p class="board__no-tasks">
+              Click «ADD NEW TASK» in menu to create your first task
+          </p>`
+    );
+};
+
+export { createNoTasksMessageTemplate };
