+++ src/view/tasks-list.js
@@ -1,0 +1,9 @@
+const createTasksListTemplate = () => {
+    return (
+        ` <div class="board__tasks">
+          </div>
+          `
+    );
+};
+
+export { createTasksListTemplate };
